requirements.txt
# AI & ML
openai
tiktoken
chromadb
rank_bm25
sentence-transformers
networkx
tf-keras

# PDF Processing
PyPDF2
pdfplumber

# Document Processing
python-docx
openpyxl
pandas

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles

# Utilities
python-dotenv
colorama

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==9.0.2
httpx==0.28.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
requests

